feat(import): allow to setup fileSizeChangeDelay
This is a new import option that can be set to a different value for
each import script.
If this option is set to a value > 0, the import process will for each
file read its size, wait the specified amount of time (in ms) and only
import it if the size has not changed. Otherwise, it will wait again
until the file size stays constant.

Closes #129